Base generic template linter
This involved refactoring the TemplateLinter into a separate package, making it
generic-ified to abstract out the vscode specified parts, moving some utilities
and unit tests with it, and then adapting it back to the vscode linting logic.
