+++ packages/analyticsdx-template-lint/package.json
@@ -1,0 +1,39 @@
+{
+  "version": "0.5.0",
+  "publisher": "salesforce",
+  "name": "analyticsdx-template-lint",
+  "displayName": "Salesforce Analytics Template Lint",
+  "description": "Provides a linting api for Salesforce/TableauCRM analytics templates",
+  "license": "BSD-3-Clause",
+  "dependencies": {
+    "fuse.js": "3.4.6",
+    "jsonc-parser": "2.1.0"
+  },
+  "devDependencies": {
+    "@types/chai": "^4.0.0",
+    "@types/mocha": "7.0.1",
+    "@types/node": "12.12.18",
+    "chai": "^4.0.2",
+    "istanbul": "^0.4.5",
+    "mocha": "7.0.1",
+    "mocha-junit-reporter": "^1.13.0",
+    "mocha-multi-reporters": "^1.1.4",
+    "nyc": "^15.0.0",
+    "remap-istanbul": "^0.9.5",
+    "shx": "0.3.2",
+    "source-map-support": "^0.5.16",
+    "typescript": "4.1.3"
+  },
+  "files": [
+    "out/src"
+  ],
+  "main": "out/src/index.js",
+  "scripts": {
+    "compile": "tsc -p ./",
+    "lint": "tslint --project .",
+    "watch": "tsc -watch -p .",
+    "clean": "shx rm -rf node_modules && shx rm -rf out && shx rm -rf coverage && shx rm -rf .nyc_output",
+    "test": "npm run test:unit",
+    "test:unit": "./node_modules/.bin/nyc ./node_modules/.bin/_mocha --recursive out/test/unit --reporter mocha-multi-reporters --timeout 20000 --reporter mocha-multi-reporters --reporter-options configFile=../../config/mochaUnitTestsConfig.json"
+  }
+}
